requirements.txt
chromadb
langchain
openai
streamlit
tiktoken
python-dotenv
